Keep networked credential proof at the integration boundary
The gateway assertion exercises a framework stream, provider resolution, and the mocked transport. Keeping that proof in a colocated unit file misclassified a multi-component network boundary and failed the semantic test audit.

Constraint: Unit tests may not acquire new network effects.

Rejected: Add the unit file to the semantic migration inventory | grows known boundary debt

Confidence: high

Scope-risk: narrow

Reversibility: clean

Directive: Keep the end-to-end Authorization assertion in the integration suite.

Tested: semantic disposition audit, Deno format/lint, integration credential test, and 66 focused run-stream steps.

Not-tested: Full CI is rerunning on the new head.
